+++ demisto_sdk/commands/common/errors.py
@@ -454,9 +454,10 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
-    def incident_in_command_name_or_args(strings_contain_incident):
+    def incident_in_command_name_or_args(commands, args):
         return f"This is a core pack with an integration that contains the word incident in the following commands'" \
-               f" name or argument:\n {strings_contain_incident}. \n, To fix the problem, remove the word incident, " \
+               f" name or argument:\ncommand's name: {commands} \ncommand's argument: {args}\n" \
+               f"To fix the problem, remove the word incident, " \
                f"or add them to the whitelist named argsExceptionsList in:\n" \
                f"https://github.com/demisto/server/blob/57fbe417ae420c41ee12a9beb850ff4672209af8/services/" \
                f"servicemodule_test.go#L8273"
@@ -965,7 +966,8 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def incident_in_script_arg(arguments):
-        return f"The script is part of a core pack. Therefore, the use of the word `incident` in argument names is forbidden. problematic argument" \
+        return f"The script is part of a core pack. Therefore, the use of the word `incident` in argument names is" \
+               f" forbidden. problematic argument" \
                f" names:\n {arguments}. \n, To fix the problem, remove the word incident, " \
                f"or add the argument name to the allowlist named argsExceptionsList in:\n" \
                f"https://github.com/demisto/server/blob/57fbe417ae420c41ee12a9beb850ff4672209af8/services/" \